requirements.txt
django==2.0.2
django-filter==1.1.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-cors-headers==2.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-role-permissions==2.2.0
djangorestframework-filters==0.10.2
django-simple-captcha==0.5.6
httpsig==1.1.2
mysqlclient==1.3.12
Pillow==5.0.0
pip==9.0.3
pycrypto==2.6.1
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
requests==2.18.4
